refactor: extract FilesCopyHandlerImpl
